Move ImmediateWindowProvider to the SPI project, leaving behind an interface for binary compatibility
